requirements.txt
Flask
twilio
python-dotenv
groq
reportlab
streamlit
python-docx
